// src/app/models/maintenance/auto-fail-config/autoFailConfig.ts
export class AutoFailConfig {
    SEQID: number;
    OEM_CD: string;
    MODEL: string;
    PROGRAM_NAME: string;
    PROGRAM_INDICATOR: string;
    CONDITION_CD: string;
    TRIGGER_LEVEL: number;
    TRIGGER_ATTR: string;
    ACTIVE: string = 'Y';
    RANK: string;
    FORMULA: string;
    AUTOFAIL_TRANSID: string;
    IVCCODE: string;
    IVC_CATEGORY_OVERRIDE: string;
    CLIENTID: string;
    SITEID: string;
    ADDTS: string;
    ADDWHO: string;
    EDITTS: string;
    EDITWHO: string;
}